used spi irq instead, and may have updated the queue stuff. This is on the 18/02/2026

+++ file_processing/file_processing.c
@@ -338,7 +338,7 @@ static void add_subdirectory() //need to get date time stamp and stuff and use t
 
 
 
-PUBLIC unsigned int append_char(uint16_t byte)
+PUBLIC unsigned int append_char(uint32_t byte)
 {
     static unsigned int i = 0;
